memset working

+++ src/my-malloc.h
@@ -15,4 +15,4 @@ void *myMalloc(uint32_t size);
 void myFree(void *ptr);
 void memoryMap(void);
 int myFreeErrorCode(void *ptr);
-int myMemset(void *p, uint8_t val, uint32_t len);
+int myMemset(void *p, long val, long len);
